+++ Microsoft.Azure.Cosmos/src/Linq/TypeSystem.cs
@@ -10,60 +10,13 @@ namespace Microsoft.Azure.Cosmos.Linq
     using System.Linq;
     using System.Reflection;
     using System.Runtime.CompilerServices;
-    using System.Runtime.Serialization;
-    using Microsoft.Azure.Cosmos.Serializer;
     using Microsoft.Azure.Documents;
-    using Newtonsoft.Json;
 
     internal static class TypeSystem
     {
         public static Type GetElementType(Type type)
         {
             return GetElementType(type, new HashSet<Type>());
-        }
-
-        public static string GetMemberName(this MemberInfo memberInfo, CosmosLinqSerializerOptions linqSerializerOptions = null)
-        {
-            string memberName = null;
-
-            // Check if Newtonsoft JsonExtensionDataAttribute is present on the member, if so, return empty member name.
-            JsonExtensionDataAttribute jsonExtensionDataAttribute = memberInfo.GetCustomAttribute<JsonExtensionDataAttribute>(true);
-            if (jsonExtensionDataAttribute != null && jsonExtensionDataAttribute.ReadData)
-            {
-                return null;
-            }
-
-            // Json.Net honors JsonPropertyAttribute more than DataMemberAttribute
-            // So we check for JsonPropertyAttribute first.
-            JsonPropertyAttribute jsonPropertyAttribute = memberInfo.GetCustomAttribute<JsonPropertyAttribute>(true);
-            if (jsonPropertyAttribute != null && !string.IsNullOrEmpty(jsonPropertyAttribute.PropertyName))
-            {
-                memberName = jsonPropertyAttribute.PropertyName;
-            }
-            else
-            {
-                DataContractAttribute dataContractAttribute = memberInfo.DeclaringType.GetCustomAttribute<DataContractAttribute>(true);
-                if (dataContractAttribute != null)
-                {
-                    DataMemberAttribute dataMemberAttribute = memberInfo.GetCustomAttribute<DataMemberAttribute>(true);
-                    if (dataMemberAttribute != null && !string.IsNullOrEmpty(dataMemberAttribute.Name))
-                    {
-                        memberName = dataMemberAttribute.Name;
-                    }
-                }
-            }
-
-            if (memberName == null)
-            {
-                memberName = memberInfo.Name;
-            }
-
-            if (linqSerializerOptions != null)
-            {
-                memberName = CosmosSerializationUtil.GetStringWithPropertyNamingPolicy(linqSerializerOptions, memberName);
-            }
-
-            return memberName;
         }
 
         private static Type GetElementType(Type type, HashSet<Type> visitedSet)
